repair_single_precision preliminary solution to eliminate compile errors when using single preciosn
Changed all the uses of double that produced compile errors to realnum.
Added macro TO_CREALNUM(a) that converts in-place from
complex<double/float> to complex<realnum>. This solution is a
workaround.

+++ src/meepgeom.cpp
@@ -2299,7 +2299,7 @@ meep::realnum get_material_gradient(
 
   std::complex<double> dA_du[9] = {std::complex<double>(0, 0)};
   for (int i = 0; i < 9; i++)
-    dA_du[i] = (dA_du_1[i] - dA_du_0[i]) / (2 * du);
+    dA_du[i] = (dA_du_1[i] - dA_du_0[i]) / (2.0 * double(du));
 
   int dir_idx;
   if (field_dir == meep::Ex)
